Extract overlay mapper lifecycle wiring
