+++ Sources/KeyPathAppKit/UI/Overlay/OverlayMapperSection.swift
@@ -36,7 +36,7 @@ struct OverlayMapperSection: View {
     /// Selected tap count (1 = single, 2 = double, 3 = triple)
     @State var selectedTapCount: Int = 1
     /// Cached set of installed pack IDs (for green dot on visual-only packs)
-    @State private var installedPackIDs: Set<String> = []
+    @State var installedPackIDs: Set<String> = []
 
     /// Current behavior slot being edited (tap is default)
     @State var selectedBehaviorSlot: BehaviorSlot = .tap
@@ -55,376 +55,6 @@ struct OverlayMapperSection: View {
 
     var body: some View {
         bodyView
-    }
-
-    private var bodyView: some View {
-        let base = VStack(spacing: 8) {
-            if shouldShowHealthGate {
-                healthGateContent
-            } else {
-                mapperContent
-            }
-        }
-
-        let withAppear = base.onAppear {
-            Task {
-                let records = await InstalledPackTracker.shared.allInstalled()
-                installedPackIDs = Set(records.map(\.packID))
-            }
-            guard !shouldShowHealthGate, let kanataViewModel else { return }
-            viewModel.configure(kanataManager: kanataViewModel.underlyingManager)
-            viewModel.setLayer(kanataViewModel.currentLayerName)
-            // Initialize with actual mapping for A key (keyCode 0)
-            // Look up from layerKeyMap to get current remapping, if any
-            let defaultKeyCode: UInt16 = 0 // 'A' key on macOS
-            let layerInfo = layerKeyMap[defaultKeyCode]
-            let inputLabel = "a"
-            let outputLabel = layerInfo?.displayLabel ?? "a"
-            let appId = layerInfo?.appLaunchIdentifier
-            let systemId = layerInfo?.systemActionIdentifier
-            let urlId = layerInfo?.urlIdentifier
-            viewModel.setInputFromKeyClick(
-                keyCode: defaultKeyCode,
-                inputLabel: inputLabel,
-                outputLabel: outputLabel,
-                appIdentifier: appId,
-                systemActionIdentifier: systemId,
-                urlIdentifier: urlId
-            )
-            // Load any existing behavior (hold action, etc.) from saved rules
-            viewModel.loadBehaviorFromExistingRule(kanataManager: kanataViewModel.underlyingManager)
-            // Notify parent to highlight the A key
-            onKeySelected?(viewModel.inputKeyCode)
-            // Initialize configured behavior slots
-            updateConfiguredBehaviorSlots()
-        }
-
-        let withPackChanges = withAppear.onReceive(
-            NotificationCenter.default.publisher(for: .installedPacksChanged)
-        ) { _ in
-            Task {
-                let records = await InstalledPackTracker.shared.allInstalled()
-                installedPackIDs = Set(records.map(\.packID))
-            }
-        }
-
-        let withDisappear = withPackChanges.onDisappear {
-            viewModel.stopKeyCapture()
-        }
-
-        let withExit = withDisappear.onExitCommand {
-            if isAnyRecordingActive {
-                cancelAllRecording()
-            }
-        }
-
-        let withInputChange = withExit.onChange(of: viewModel.inputKeyCode) { _, newKeyCode in
-            onKeySelected?(newKeyCode)
-            // Reset to tap slot when selecting a new key
-            selectedBehaviorSlot = .tap
-            selectedTapOutputMode = .default
-            selectedTapCount = 1
-            updateConfiguredBehaviorSlots()
-        }
-
-        let withLayerChange = withInputChange.onReceive(
-            NotificationCenter.default.publisher(for: .kanataLayerChanged)
-        ) { notification in
-            if let layerName = notification.userInfo?["layerName"] as? String {
-                viewModel.setLayer(layerName)
-            }
-        }
-
-        let withDrawerSelection = withLayerChange.onReceive(
-            NotificationCenter.default.publisher(for: .mapperDrawerKeySelected)
-        ) { notification in
-            guard !shouldShowHealthGate else { return }
-            guard let keyCode = notification.userInfo?["keyCode"] as? UInt16,
-                  let inputKey = notification.userInfo?["inputKey"] as? String,
-                  let outputKey = notification.userInfo?["outputKey"] as? String
-            else { return }
-
-            // Extract optional action identifiers
-            let appId = notification.userInfo?["appIdentifier"] as? String
-            let systemId = notification.userInfo?["systemActionIdentifier"] as? String
-            let urlId = notification.userInfo?["urlIdentifier"] as? String
-            let shiftedOutputKey = notification.userInfo?["shiftedOutputKey"] as? String
-            let displayLabel = notification.userInfo?["displayLabel"] as? String
-
-            // Extract app condition info (for editing app-specific rules)
-            let appBundleId = notification.userInfo?["appBundleId"] as? String
-            let appDisplayName = notification.userInfo?["appDisplayName"] as? String
-
-            // Update the mapper's input to the clicked key
-            viewModel.setInputFromKeyClick(
-                keyCode: keyCode,
-                inputLabel: inputKey,
-                outputLabel: outputKey,
-                appIdentifier: appId,
-                systemActionIdentifier: systemId,
-                urlIdentifier: urlId,
-                shiftedOutputKey: shiftedOutputKey
-            )
-            // Override display label when it differs from outputKey (e.g., Hyper shows "✦" not "lctl")
-            if let displayLabel, displayLabel != outputKey {
-                viewModel.outputLabel = viewModel.formatKeyForDisplay(displayLabel)
-            }
-            selectedTapOutputMode = shiftedOutputKey == nil ? .default : .shifted
-
-            // Load any existing behavior (hold action, etc.) from saved rules
-            if let manager = kanataViewModel?.underlyingManager {
-                viewModel.loadBehaviorFromExistingRule(kanataManager: manager)
-            }
-
-            // If app condition is provided, set it for app-specific rule editing
-            if let bundleId = appBundleId, let displayName = appDisplayName {
-                // Load the app icon (or use fallback)
-                let icon: NSImage
-                if let appURL = NSWorkspace.shared.urlForApplication(withBundleIdentifier: bundleId) {
-                    icon = NSWorkspace.shared.icon(forFile: appURL.path)
-                    icon.size = NSSize(width: 24, height: 24)
-                } else {
-                    icon = NSImage(systemSymbolName: "app.fill", accessibilityDescription: nil)
-                        ?? NSImage()
-                }
-                viewModel.selectedAppCondition = AppConditionInfo(
-                    bundleIdentifier: bundleId,
-                    displayName: displayName,
-                    icon: icon
-                )
-            }
-        }
-
-        let withOpenAppConditionPicker = withDrawerSelection.onReceive(
-            NotificationCenter.default.publisher(for: .openMapperAppConditionPicker)
-        ) { _ in
-            // Open the app condition picker when requested (e.g., from "New Rule" button in App Rules tab)
-            guard !shouldShowHealthGate else { return }
-            // Small delay to allow tab switch animation to complete
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.15) {
-                isAppConditionPickerOpen = true
-            }
-        }
-
-        let withSetAppCondition = withOpenAppConditionPicker.onReceive(
-            NotificationCenter.default.publisher(for: .mapperSetAppCondition)
-        ) { notification in
-            // Set app condition directly (from "+" button on app card in App Rules tab)
-            guard !shouldShowHealthGate else { return }
-            guard let bundleId = notification.userInfo?["bundleId"] as? String,
-                  let displayName = notification.userInfo?["displayName"] as? String
-            else { return }
-
-            // Load app icon
-            let icon: NSImage
-            if let appURL = NSWorkspace.shared.urlForApplication(withBundleIdentifier: bundleId) {
-                icon = NSWorkspace.shared.icon(forFile: appURL.path)
-                icon.size = NSSize(width: 24, height: 24)
-            } else {
-                icon = NSImage(systemSymbolName: "app.fill", accessibilityDescription: nil) ?? NSImage()
-            }
-
-            viewModel.selectedAppCondition = AppConditionInfo(
-                bundleIdentifier: bundleId,
-                displayName: displayName,
-                icon: icon
-            )
-
-            // Reset input/output for new rule
-            viewModel.resetForNewMapping()
-        }
-
-        let withAutoSave = withSetAppCondition.onChange(of: viewModel.isRecordingOutput) { wasRecording, isRecording in
-            // Only trigger when recording just stopped
-            guard wasRecording, !isRecording else { return }
-            guard !shouldShowHealthGate else { return }
-
-            // Check if we have a valid mapping (input != output or has an action)
-            let hasOutputAction = viewModel.selectedApp != nil ||
-                viewModel.selectedSystemAction != nil ||
-                viewModel.selectedURL != nil
-            let hasKeyRemapping = viewModel.inputLabel.lowercased() != viewModel.outputLabel.lowercased()
-
-            guard hasOutputAction || hasKeyRemapping else { return }
-            guard let manager = kanataViewModel?.underlyingManager else { return }
-
-            // Auto-save the mapping
-            Task {
-                await viewModel.save(kanataManager: manager)
-                // Update configured slots after save
-                updateConfiguredBehaviorSlots()
-            }
-        }
-
-        let withShiftAutoSave = withAutoSave.onChange(of: viewModel.isRecordingShiftedOutput) { wasRecording, isRecording in
-            guard wasRecording, !isRecording else { return }
-            guard !shouldShowHealthGate else { return }
-            guard viewModel.hasShiftedOutputConfigured else { return }
-            guard let manager = kanataViewModel?.underlyingManager else { return }
-            Task {
-                await viewModel.save(kanataManager: manager)
-                updateConfiguredBehaviorSlots()
-            }
-        }
-
-        let withDoubleTapAutoSave = withShiftAutoSave.onChange(of: viewModel.isRecordingDoubleTap) { wasRecording, isRecording in
-            guard wasRecording, !isRecording else { return }
-            guard !shouldShowHealthGate else { return }
-            guard !viewModel.doubleTapAction.isEmpty else { return }
-            guard let manager = kanataViewModel?.underlyingManager else { return }
-            Task {
-                await viewModel.save(kanataManager: manager)
-                updateConfiguredBehaviorSlots()
-            }
-        }
-
-        let withHoldAutoSave = withDoubleTapAutoSave.onChange(of: viewModel.isRecordingHold) { wasRecording, isRecording in
-            guard wasRecording, !isRecording else { return }
-            guard !shouldShowHealthGate else { return }
-            guard !viewModel.holdAction.isEmpty else { return }
-            guard let manager = kanataViewModel?.underlyingManager else { return }
-            Task {
-                await viewModel.save(kanataManager: manager)
-                updateConfiguredBehaviorSlots()
-            }
-        }
-
-        let withOutputChange = withHoldAutoSave.onChange(of: viewModel.outputLabel) { _, _ in
-            updateConfiguredBehaviorSlots()
-        }
-
-        let withShiftedOutputChange = withOutputChange.onChange(of: viewModel.shiftedOutputLabel) { _, _ in
-            updateConfiguredBehaviorSlots()
-        }
-
-        let withHoldChange = withShiftedOutputChange.onChange(of: viewModel.holdAction) { _, _ in
-            updateConfiguredBehaviorSlots()
-        }
-
-        let withDoubleTapChange = withHoldChange.onChange(of: viewModel.doubleTapAction) { _, _ in
-            updateConfiguredBehaviorSlots()
-        }
-
-        let withTapDanceChange = withDoubleTapChange.onChange(of: viewModel.tapDanceSteps.map(\.action)) { _, _ in
-            updateConfiguredBehaviorSlots()
-        }
-
-        let withAppChange = withTapDanceChange.onChange(of: viewModel.selectedApp?.name) { _, _ in
-            updateConfiguredBehaviorSlots()
-        }
-
-        let withSystemChange = withAppChange.onChange(of: viewModel.selectedSystemAction?.id) { _, _ in
-            updateConfiguredBehaviorSlots()
-        }
-
-        let withShiftAvailability = withSystemChange.onChange(of: viewModel.canUseShiftedOutput) { _, canUse in
-            if !canUse {
-                if selectedTapOutputMode == .shifted {
-                    selectedTapOutputMode = .default
-                }
-                if selectedBehaviorSlot == .shift {
-                    selectedBehaviorSlot = .tap
-                }
-            }
-        }
-
-        let withSlotChange = withShiftAvailability.onChange(of: selectedBehaviorSlot) { _, newSlot in
-            selectedTapOutputMode = newSlot == .shift ? .shifted : .default
-            if newSlot != .tap { selectedTapCount = 1 }
-            playBehaviorAnimation(for: newSlot)
-        }
-
-        let withTapModeChange = withSlotChange.onChange(of: selectedTapOutputMode) { _, newMode in
-            if newMode == .shifted, viewModel.isRecordingOutput {
-                viewModel.stopRecording()
-            } else if newMode == .default, viewModel.isRecordingShiftedOutput {
-                viewModel.stopRecording()
-            }
-        }
-
-        let withTapCountChange = withTapModeChange.onChange(of: selectedTapCount) { _, newCount in
-            if newCount > 1 {
-                // Show label animation for multi-tap
-                playBehaviorAnimation(for: .tap)
-            }
-        }
-
-        let withResetDialog = withTapCountChange.confirmationDialog(
-            "Clear Mapping",
-            isPresented: $showingResetDialog,
-            titleVisibility: .visible
-        ) {
-            // Option 1: Clear just the current slot (only for non-tap slots that are configured)
-            if selectedBehaviorSlot != .tap, currentSlotIsConfigured {
-                Button("Clear \(selectedBehaviorSlot.label) Only") {
-                    clearCurrentSlot()
-                }
-                .accessibilityIdentifier("overlay-mapper-reset-slot-button")
-            }
-
-            // Option 2: Clear all behaviors for this key
-            Button("Clear All for \"\(viewModel.inputLabel.uppercased())\"") {
-                clearAllBehaviorsForCurrentKey()
-            }
-            .accessibilityIdentifier("overlay-mapper-reset-key-button")
-
-            // Option 3: Clear everything
-            Button("Clear All Custom Mappings", role: .destructive) {
-                performResetAll()
-            }
-            .accessibilityIdentifier("overlay-mapper-reset-all-button")
-
-            Button("Cancel", role: .cancel) {}
-                .accessibilityIdentifier("overlay-mapper-reset-cancel-button")
-        } message: {
-            if selectedBehaviorSlot != .tap, currentSlotIsConfigured {
-                Text("What would you like to clear?")
-            } else {
-                Text("Choose what to clear for \"\(viewModel.inputLabel.uppercased())\"")
-            }
-        }
-
-        let selectedKeyInfo: LayerKeyInfo? = viewModel.inputKeyCode.flatMap { layerKeyMap[$0] }
-        let withLayerMapRefresh = withResetDialog.onChange(of: selectedKeyInfo) { _, newInfo in
-            guard let keyCode = viewModel.inputKeyCode, let newInfo else { return }
-            let inputKey = OverlayKeyboardView.keyCodeToKanataName(keyCode)
-            let outputKey = newInfo.outputKey ?? newInfo.displayLabel
-            guard !outputKey.isEmpty else { return }
-            viewModel.setInputFromKeyClick(
-                keyCode: keyCode,
-                inputLabel: inputKey,
-                outputLabel: outputKey,
-                appIdentifier: newInfo.appLaunchIdentifier,
-                systemActionIdentifier: newInfo.systemActionIdentifier,
-                urlIdentifier: newInfo.urlIdentifier
-            )
-            if !newInfo.displayLabel.isEmpty, newInfo.displayLabel != outputKey {
-                viewModel.outputLabel = viewModel.formatKeyForDisplay(newInfo.displayLabel)
-            }
-            if let manager = kanataViewModel?.underlyingManager {
-                viewModel.loadBehaviorFromExistingRule(kanataManager: manager)
-            }
-        }
-
-        let withURLSheet = withLayerMapRefresh.sheet(isPresented: $viewModel.showingURLDialog) {
-            URLInputDialog(
-                urlText: $viewModel.urlInputText,
-                onSubmit: { viewModel.submitURL() },
-                onCancel: { viewModel.showingURLDialog = false }
-            )
-        }
-
-        return withURLSheet.sheet(isPresented: $showingAppPickerSheet) {
-            let vm = viewModel
-            AppConditionPickerSheet(
-                onSelect: { condition in
-                    vm.selectedAppCondition = condition
-                },
-                onBrowse: {
-                    pickAppForCondition()
-                }
-            )
-        }
     }
 
     private let showDebugBorders = false
@@ -476,7 +106,7 @@ struct OverlayMapperSection: View {
         }
     }
 
-    private var mapperContent: some View {
+    var mapperContent: some View {
         SlideOverContainer(
             isPresented: $showMultiTapSlideOver,
             panelTitle: "Multi-tap Actions",
